New connections now handled in background
Server now checks for new connections in a thread and then creates a new thread to deal with whether they create or join a room.

Also, now that all happens in the background while other code is running (here, filler code that outputs the filled and unfilled rooms is what runs).